feat: scroll mobile availability list to today
