+++ src/pages/availability/week/views/mobile-view/AvailabilityWeekMobileList.types.ts
@@ -1,0 +1,16 @@
+import type {
+	IAvailabilityWeekMobileDay,
+	IWeekSlot,
+} from '../../AvailabilityWeekPage.types';
+
+export type AvailabilityWeekMobileListProps = {
+	days: IAvailabilityWeekMobileDay[];
+	onDayHeaderClick: (
+		event: React.MouseEvent<HTMLElement>,
+		dateStr: string
+	) => void;
+	onSlotClick: (slot: IWeekSlot) => void;
+	onSlotPointerDown: (slot: IWeekSlot) => void;
+	onToggleDayExpanded: (dateStr: string) => void;
+	todayCardId: string;
+};
